feat: Refactor task deletion logic and error handling

+++ Task-6/task_manager/controllers/user_controller.go
@@ -3,21 +3,24 @@ package controllers
 import (
 	"net/http"
 	"os"
+	"time"
 
 	"github.com/Nahom-Derese/Learning_Go/Task-6/task_manager/data"
 	"github.com/Nahom-Derese/Learning_Go/Task-6/task_manager/models"
 	"github.com/gin-gonic/gin"
-	"github.com/golang-jwt/jwt"
+	"github.com/golang-jwt/jwt/v5"
 	"golang.org/x/crypto/bcrypt"
 )
 
 type UserHandlers interface {
 	GetAllUsers() gin.HandlerFunc
+	DeleteAllUsers() gin.HandlerFunc
 	GetUserByUsername() gin.HandlerFunc
 	UpdateUser() gin.HandlerFunc
 	DeleteUser() gin.HandlerFunc
-	CreateUser() gin.HandlerFunc
+	PromoteUser() gin.HandlerFunc
 
+	SignUp() gin.HandlerFunc
 	Login() gin.HandlerFunc
 }
 
@@ -27,8 +30,19 @@ type UserController struct {
 
 func (ctrl *UserController) GetAllUsers() gin.HandlerFunc {
 	return func(c *gin.Context) {
-		tasks := ctrl.UserRepo.FindAll()
-		c.JSON(200, tasks)
+		users := ctrl.UserRepo.FindAll()
+		c.JSON(200, users)
+	}
+}
+
+func (ctrl *UserController) DeleteAllUsers() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		err := ctrl.UserRepo.DeleteAll()
+		if err != nil {
+			c.JSON(http.StatusNotFound, gin.H{"error": "users not found"})
+			return
+		}
+		c.JSON(http.StatusAccepted, gin.H{"message": "users deleted"})
 	}
 }
 
@@ -36,6 +50,13 @@ func (ctrl *UserController) GetUserByUsername() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		username := c.Param("username")
 		user, err := ctrl.UserRepo.FindByUsername(username)
+
+		userData, _ := c.Get("user")
+		if username != userData.(*models.User).Username {
+			c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
+			return
+		}
+
 		if err != nil {
 			c.JSON(http.StatusNotFound, gin.H{"error": "user not found"})
 			return
@@ -48,7 +69,14 @@ func (ctrl *UserController) UpdateUser() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		username := c.Param("username")
 
-		user, err := ctrl.UserRepo.FindByUsername(username)
+		userData, _ := c.Get("user")
+
+		if username != userData.(*models.User).Username {
+			c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
+			return
+		}
+
+		user, err := ctrl.UserRepo.FindUser(username)
 		if err != nil {
 			c.JSON(http.StatusNotFound, gin.H{"error": "user not found"})
 			return
@@ -65,9 +93,44 @@ func (ctrl *UserController) UpdateUser() gin.HandlerFunc {
 	}
 }
 
+func (ctrl *UserController) PromoteUser() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		username := c.Param("username")
+
+		user, err := ctrl.UserRepo.FindUser(username)
+		if err != nil {
+			c.JSON(http.StatusNotFound, gin.H{"error": "user not found"})
+			return
+		}
+		c.BindJSON(&user)
+
+		if user.Role == "admin" {
+			c.JSON(http.StatusConflict, gin.H{"error": "user is already an admin"})
+			return
+		}
+
+		user.Role = "admin"
+		newUser, err := ctrl.UserRepo.Save(user)
+
+		if err != nil {
+			c.JSON(http.StatusConflict, gin.H{"error": err.Error()})
+			return
+		}
+
+		c.JSON(http.StatusOK, newUser)
+	}
+}
+
 func (ctrl *UserController) DeleteUser() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		username := c.Param("username")
+
+		userData, _ := c.Get("user")
+
+		if username != userData.(*models.User).Username {
+			c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
+			return
+		}
 
 		err := ctrl.UserRepo.Delete(username)
 		if err != nil {
@@ -78,13 +141,9 @@ func (ctrl *UserController) DeleteUser() gin.HandlerFunc {
 	}
 }
 
-func (ctrl *UserController) CreateUser() gin.HandlerFunc {
+func (ctrl *UserController) SignUp() gin.HandlerFunc {
 	return func(c *gin.Context) {
-		var user struct {
-			Username string `json:"username"`
-			Password string `json:"password"`
-			Role     string `json:"role"`
-		}
+		var user models.User
 
 		err := c.BindJSON(&user)
 
@@ -100,10 +159,20 @@ func (ctrl *UserController) CreateUser() gin.HandlerFunc {
 			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		}
 
-		newUser := models.User{Username: user.Username, Password: string(hash), Role: user.Role}
+		// if no users in the db, create an admin user
+		users := ctrl.UserRepo.FindAll()
 
-		ctrl.UserRepo.Save(&newUser)
-		c.JSON(200, user)
+		role := "regular"
+		if len(users) == 0 {
+			role = "admin"
+		}
+
+		newUser := models.User{Username: user.Username, Password: string(hash), Role: role}
+		ctrl.UserRepo.Save(newUser)
+
+		user.Role = role
+
+		c.JSON(http.StatusOK, user)
 
 	}
 }
@@ -122,7 +191,7 @@ func (ctrl *UserController) Login() gin.HandlerFunc {
 		}
 
 		// get user from db
-		userFromDB, err := ctrl.UserRepo.FindByUsername(user.Username)
+		userFromDB, err := ctrl.UserRepo.FindUser(user.Username)
 		if err != nil {
 			c.JSON(http.StatusNotFound, gin.H{"error": "user not found"})
 			return
@@ -132,6 +201,7 @@ func (ctrl *UserController) Login() gin.HandlerFunc {
 
 		if err != nil {
 			c.JSON(http.StatusUnauthorized, gin.H{"error": "invalid password"})
+			return
 		}
 
 		// generate token
@@ -139,6 +209,7 @@ func (ctrl *UserController) Login() gin.HandlerFunc {
 
 		if err != nil {
 			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			return
 		}
 
 		c.JSON(http.StatusOK, gin.H{"token": token})
@@ -151,7 +222,7 @@ func GenerateToken(username, role string) (string, error) {
 	token, err := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
 		"username": username,
 		"role":     role,
-		"exp":      jwt.TimeFunc().AddDate(0, 0, 1).Unix(),
+		"exp":      time.Now().AddDate(0, 0, 1).Unix(),
 	}).SignedString([]byte(os.Getenv("JWT_SECRET")))
 
 	if err != nil {
